refactor(#3): rename folders
