+++ Makefile.vars
@@ -3,6 +3,6 @@ BINARY_NAME	:= bragdoc
 GOPATH		:= $(shell go env GOPATH)
 VERSION		:=$(shell git describe --abbrev=0 --tags 2> /dev/null || echo "0.1.0")
 BUILD		:=$(shell git rev-parse --short HEAD 2> /dev/null || echo "undefined")
-LDFLAGS		:="-X 'github.com/vagnerclementino/bragdoc/internal/cmd.Version=$(VERSION)' -X 'github.com/vagnerclementino/bragdoc/internal/cmd.Build=$(BUILD)'"
+LDFLAGS		:="-X 'github.com/vagnerclementino/bragdoc/internal/handler.Version=$(VERSION)' -X 'github.com/vagnerclementino/bragdoc/internal/handler.Build=$(BUILD)'"
 GOOS		:=darwin
 GOARCH		:=amd64
