+++ go.mod
@@ -4,6 +4,7 @@ go 1.21.1
 
 require (
 	github.com/google/uuid v1.3.1
+	github.com/spf13/cobra v1.7.0
 	github.com/stretchr/testify v1.8.4
 )
 
@@ -11,7 +12,6 @@ require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/spf13/cobra v1.7.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
