--- main.go
@@ -1,11 +1,0 @@
-/*
-Copyright © 2023 Vagner Clementino vagner.clementino@gmail.com
-
-*/
-package main
-
-import "github.com/vagnerclementino/bragdoc/cmd"
-
-func main() {
-	cmd.Execute()
-}
